refactor: fixed typo in shurjopay_plugin, changed installation guideline in read
