+++ src/shurjopay_plugin/logger_config.py
